split tests in three different classes that depend on each other
Also handled flat issue with SMILE and documented stuff
